+++ src/Components/CollaboratorCard/index.js
@@ -1,54 +1,22 @@
-import React, { useContext, useState, useEffect } from 'react';
+import React from 'react';
 import { FaEdit } from 'react-icons/all';
-import { AuthContext } from '../../Contexts/AuthContext';
-import CollaboratorService from '../../Services/CollaboratorService'
 import './styles.css';
 
-export default function Footer() {
-  const { user } = useContext(AuthContext);
-  const [collaborators, setCollaborators] = useState([]);
-
-  useEffect(() => {
-    if(user) loadCollaborators();
-  }, []);
-  
-  const loadCollaborators = async () => {
-    try {
-      const response = await CollaboratorService.loadCollaborator(user.idCompany);
-      console.log(response);
-      setCollaborators(response);
-
-    } catch(error) {
-      console.log(error);
-    }
-  }
-  const nameAccessLevel = (user) => {
-    let accessLevel;
-    switch (user.idAccessLevel) {
-      case 1:
-        accessLevel = "Gestor";
-        break;
-      case 2:
-        accessLevel = "Administrador";
-        break;
-      case 3:
-        accessLevel = "Vendedor";
-        break;
-      default:
-    }
-    return accessLevel;
-  }
-
+export default function CollaboratorCard(props) {
   return (
     <div class="col-md-6">
       <div className="container">
         <div className="card-header">
-          <img src='#' />
+          <img src={props.photo} />
           <div className="collaborator">
-            <p className="name">{user.name}</p>
-            <p className="access-level">{nameAccessLevel(user)}</p>
+            <p className="name">{props.name}</p>
+            <p className="access-level">{props.accessLevel}</p>
           </div>
           <FaEdit size={25} />
+        </div>
+        <div className="collaborator-actions">
+          <p className="deactivate">Desativar Colaborador</p>
+          <p className="show-sells">Visualizar Vendas</p>
         </div>
       </div>
     </div>
